add the Spotlight Component
