+++ pages/_app.js
@@ -1,10 +1,24 @@
 import GlobalStyle from "../styles";
+import useSWR from "swr";
+import { useRouter } from "next/router";
+import Image from "next/image";
+import ArtPieces from "../components/ArtPieces";
+
+const fetcher = (URL) => fetch(URL).then((res) => res.json());
 
 export default function App({ Component, pageProps }) {
+  const { data: pieces } = useSWR(
+    "https://example-apis.vercel.app/api/art",
+    fetcher
+  );
+
+  if (!pieces) {
+    return "loading...";
+  }
   return (
     <>
       <GlobalStyle />
-      <Component {...pageProps} />
+      <Component {...pageProps} pieces={pieces} />
     </>
   );
 }
